add jniLibs & frameworks
