+++ mars/xlog/flutter_xlog/example/android/app/proguard-rules.pro
@@ -1,0 +1,3 @@
+
+## XLog
+-keep class com.tencent.mars.** {*;}
