+++ mars/xlog/flutter_xlog/ios/.gitignore
@@ -36,4 +36,3 @@ Icon?
 /Flutter/Generated.xcconfig
 /Flutter/ephemeral/
 /Flutter/flutter_export_environment.sh
-Frameworks/**
